PINIO BOX: Support configurable pulse on box activation
By default, PINIO BOX facility is capable of turning on the configured PINIO when the associated box/mode is active and turning it off when it is no longer active. This PR adds an ability to optionally send a pulse of a specified duration when the associated box/mode gets activated. The intended use-case for this feature is to control devices that use push-buttons which must be pressed for a certain duration to activate their functions.
An example that is presented at the end shows how this new feature can be used to control SMO 4K camera's power and recording in a more robust way than is possible via a regular PINIO BOX alone.

This PR introduces a new configuration option: "pinio_box_pulse". It is an array of four numbers, similarly to "pinio_config" and "pinio_box" options. The default value of "pinio_box_pulse" is "0,0,0,0". Zero value means that the corresponding PINIO works as usual -- it tracks the status of the corresponding box/mode as specified in "pinio_box". A non-zero value of "pinio_box_pulse" specifies a pulse duration in units of 100ms (e.g. 10 means a 1 second pulse). When the pulse is configured, then the PINIO is pulsed for the specified duration every time when the associated mode/box is activated.

Example

For example, lets take an iFlight Beast F745 AIO FC (IFLIGHT_F745_AIO) and SMO 4K camera. Let's solder camera's power control wire to T7 pad on FC and camera's record control wire to R7 pad on FC, and reconfigure the corresponding pads for PINIO:

resource SERIAL_TX 7 NONE # was set to E08
resource SERIAL_RX 7 NONE # was set to E07
resource PINIO 1 E07      # R7 pad on FC
resource PINIO 2 E08      # T7 pad on FC

Now we configure PINIO_BOX to associate USER1 and USER2 boxes with PINIO 1 and PINIO 2.

set pinio_box = 40,41,255,255

Our goal is to turn on the power on SMO 4K and turn the recording on/off, which can be robustly done with 0.5 second pulses (pressing these buttons for too long activates their alternate functions and pressing them for too short does not get reliably recognized by the camera). So, we configure "pinio_box_pulse" for both of those pins:

set pinio_box_pulse = 5,5,0,0

Now we can go to "Modes" tab in configurator and configure USER1 (Camera recording) and USER2 (Camera power) modes. For example, we can take any AUX channel and assign a full range of this channel to USER2 box. This way, USER2 becomes active as soon as FC boots and, because of "pinio_box_pulse" configuration, 0.5 second pulse is sent to the corresponding pad, powering on the attached camera. So, now the camera is powered on every time we attach a battery to FC (and will naturally be powered off when we detach the battery).

For recording control, for example, assume that we are using a high range of a three-position switch on AUX2 for PREARM. Let's set a middle range of this AUX2 switch to USER1. This way, every time we move that switch to prearm, it will activate USER1 box when the switch goes through its middle range, sending a 0.5 second pulse to the camera to turn on the recording. Accordinngly, when we design a quad, and go on to deactivate prearm swith, it goes through a middle range again, sending 0.5 second pulse to the camera which turns the recording off. It works quite reliably in practice, regardless of whether the switch is flipped fast or slow, as long as it is not flipped too fast. It needs to linger in the middle rangle for at least 20 ms to be reliably detected by PINIO BOX and trigger pulse to the camera.